matrix game with network support
